add loopback.getCurrentContext to loopback.test.js
